HACK: npcx/gpio: Disable input buffer of non-ISR GPIO and I2C
In the CL:1098864, it disabled the input buffer to bypass the excessive
power consumption on low-voltage IOs (1.8V).

This CL further disables the input buffer of GPIO/I2C, no matter it is
1.8V or 3.3V as below:
1. Disable GPIOs without ISR and not i2c ports in gpio_pre_init().
2. Disable input buffer of i2c ports before entering deep sleep.
3. Enable input buffer of i2c ports after ec wakes up.

Please note that this CL doesn't intend to fix any bug of npcx ec IOs.
It is the workwround to debug the power leakage which may caused by the
improper hardware/firmware configuration on IOs.

BRANCH=none
BUG=117139495
TEST=No build errors for "make buildall".
TEST=run stress test:
1. enter/exit S0ix
2. i2cscan after exiting deep idle.

Change-Id: I3d8a956a110c2f9c5009a3aabba24788065cc654
Signed-off-by: CHLin <CHLIN56@nuvoton.com>