Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Resolve the merge conflict first, I'll take a look once Gerrit shows "Active."  (Or if I should ignore the "Merge Conflict" let me know why, I'm still new to git & gerrit.)

I don't see any merge conflicts, but this is besides the point, this is a branch cherry-pick, you don't have to review it again. 

You can see branch in the left panel on gerrit,and in the message subject  it shows branch as well when it is not master (chromiumos/platform/ec : firmware-cr50-9308.B)

It's a shame gerrit is so noisy.

Patch-set: 1
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>