npcx: wov: Dummy out alt function defines for NPCX_ALT_SPIP_
In order to configure DEVALTN registers for alternate functions, the
NPCX_AL_TABLE is searched from first to last entry until it finds a
match for GPIO port and bit. The pin for I2S sync (GPIOA5) is also
used for SPIP_CS1. Therefore, when enabling the alternate function for
I2S_SYNC, it would match SPIP_CSI causing bit 0 of DEVALT0 to be
set. This then conflicts with bit 3 of DEVALT0 if GPIO95 needs to used
as a GPIO signal.

This CL dummies out the defines for NPCX_ALT_SPIP_ when
NPCX_WOV_SUPPORT is defined so that (A,5) will match the
NPCX_ALT_I2S_SYNC map entry instead.

BUG=b:116766596
BRANCH=none
TEST=On cheza verified that signal on GPIO95 no longer gets set low
when gpio_config_module(MODULE_WOV, 1); gets called. Also verfied that
DEVALTE has bits 1:0 set following this call.

Change-Id: Iebf4a75d68fe505a0828d774dfe13c1ec780e923
Signed-off-by: Scott Collyer <scollyer@google.com>