Update patch set 3
Patch Set 3:

This doesn't seen to actually fix the issue for the bug attached. It does seem to help another issue I was see with connecting to a DELL external monitor.

What if we kept this change but reduced the max wait time back down to 25ms (like it was before) but still poll it every 5ms?

Patch-set: 3
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>