Update patch set 13
Patch Set 13:

(20 comments)

Despite what some of the comments may suggest, this was *not* a thorough review of usb_prl_sm.c

Some of the comments may warrant checking other parts of the file for similar issues, and there may be entire classes of things to note that I didn't touch upon (mainly because it was hard to keep my eyes from glazing over with all the state machine details).

That having been said, the fact that a large portion of this code is self-explanatory reinforces my excitement that this new stack will be fantastic to work with moving forward!

Patch-set: 13
Reviewer: Gerrit User 1002022 <1002022@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>