util/flash_ec: Restore servo setting properly on error
Servo settings are not completely restored on error scenario. As an
example when flash_ec has "Failed to determine chip size", it dies
immediately. The cleanup procedure does not restore some of the servo
settings like disabling GP mode. This in turn causes EC to never
come out of reset until either the servo is disconnected and reconnected
or the servod is killed and restarted.

Restore the servo settings properly under error scenario so that EC
comes out of reset.

BRANCH=none
BUG=b:117414172
TEST=Run flash_ec and ensure that EC comes out of reset

Change-Id: I5512a44b5656ccc97686324a4d9b1bbcf07c6303
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>