anx74xx: Make ML0/ML1 unconnected if USB is enabled but DP isn't
If USB is enabled, the original logic uses the pin assignment D
which also connects the ML0 and ML1 switches, not matter DP is
enabled or not. It affects some DP signal integrity on the other
port which shares the same DP source. Should make the ML0/ML1
switches unconnected if DP is not enabled.

BRANCH=none
BUG=b:116879483
TEST=According to the bug, manually typing the i2c command like this
change to disconnect ML0/ML1 fixes the issue.

Change-Id: Icfa10b97331be7935bb626771b41eb509fa61f50
Signed-off-by: Wai-Hong Tam <waihong@google.com>