Update patch set 9
Patch Set 9:

(1 comment)

could you also rebase to top-of-tree?  looks like we now need to define
DUMMY_GPIO_BANK and fix software_panin() to compile.

could you also add an ISH FW tracking bug?  we have a master ISH
tracking bug with a few bugs hanging off of it, but it doesn't
look like we have one for the core ISH firmware.

we should probably close on the discussion we started last week
about whether we need to do a sandbox branch (or similar) for
this work.

Patch-set: 9
CC: Gerrit User 1000034 <1000034@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>