pd_protocol: Don't clear PD flags while debouncing.
The PD_FLAGS_TRY_SRC flag was being cleared every time we entered the
*_DISCONNECTED state.  However, this would lead to a case where if the
state machine was debouncing the CC lines and decided to re-enter the
SRC_DISCONNECTED state, the Try.Src flag would be cleared and the state
machine would not transition to the TryWait.SNK state after timing out.

We shouldn't clear any flags when transitioning back to the disconnected
state from the debounce state as the two states here are really the same
states in the state diagram.  This commit simply only clears the PD
flags when we're transitioning to the disconnected state but not from a
debounce state.  This also keeps the Try.Src flag set if the previous
state was a debounce as it means the state machine decided it didn't
meet the condition to exit and should continue waiting before
transitioning to TryWait.SNK.

BUG=b:115452695
BRANCH=master
TEST=Flash nocturne; boot to S0, plug in Apple 87W USB-C charger with
eMarked cable, verify we form an explicit 45W contract.

Change-Id: I6d8f5d69b8bd0d25ac7af008bbbe91f2658cdfe2
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/1286299
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>
(cherry picked from commit 1f30c7483fa5621e9d67c5977709dce73f31a66d)