pd_protocol: Keep Try.Src flag while debouncing.
The PD_FLAGS_TRY_SRC flag was being cleared every time we entered the
*_DISCONNECTED state.  However, this would lead to a case where if the
state machine was debouncing the CC lines and decided to re-enter the
SRC_DISCONNECTED state, the Try.Src flag would be cleared and the state
machine would not transition to the TryWait.SNK state after timing out.
This commit simply keeps the Try.Src flag set if the previous state was
a debounce as it means the state machine decided it didn't meet the
condition to exit and should continue waiting before transitioning to
TryWait.SNK.

BUG=b:115452695
BRANCH=firmware-nocturne-10984.B
TEST=Flash nocturne; boot to S0, plug in Apple 87W USB-C charger with
eMarked cable, verify we form an explicit 45W contract.

Change-Id: I6d8f5d69b8bd0d25ac7af008bbbe91f2658cdfe2
Signed-off-by: Aseda Aboagye <aaboagye@google.com>