Update patch set 5
Patch Set 5:

Please see if CL:1283970 works. The power on reset state of the MUX is to connect the pins. I don't know that we need to explicitly change that. The cache may be getting out of data however, so we should be able to read it directly from the device instead

Patch-set: 5
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>