cheza: Do S0->S5->S0 transition after warm_reset-toggling finished
The warm_reset is used for programming the AP SPI flash. This signal
is async that the SPMI transaction between PMIC and AP may get
wedged. So after the warm_reset-toggling is finished, should cold-
reset PMIC and AP to get a clean state.

BRANCH=None
BUG=b:112723105, b:112564635
TEST=Toggled the warm_reset on different cases, e.g. S5, S0,
coreboot, userspace, etc., and checked the AP reboot as expected,
expected S5.

Change-Id: I8d5e8690f5a6387d43f79718a8e68b2c810f4d26
Signed-off-by: Wai-Hong Tam <waihong@google.com>