ppc: syv682x: Add trigger for BC 1.2 when turning src path on/off
When using the syv682x PPC for VBUS detection, VBUS changes need to
result in a call to usb_charger_vbus_change() which will set an event
to wake the USB_CHG_N task. This was being handled when the port was
attached as a sink, but was not being done when the port was attached
as source.

This CL adds a call to usb_charger_vbus_change() when the source path
is enabled or disabled.

BRANCH=none
BUG=b:113267982
TEST=On DragonEgg, connect a USB 2.0 keyboard and verfied that it
enumerates.

localhost ~ # lsusb
Bus 003 Device 003: ID 046d:c31c Logitech, Inc. Keyboard K120

Change-Id: Ie26ad1a062320eb1290fe8c6f6f8b3de8261a03f
Signed-off-by: Scott Collyer <scollyer@google.com>