cr50: Allow SN bits to be written after BoardID.Flags was set.
While we only want SN Bits to be writable in factory, we also
want factories to se flags such as whether a board is pre-MP
as early in the process as possible, before the BoardID.Type is
available. This ensures that boards that leave the factory before
the BoardID.Type is set can still be marked as developmental boards
etc. This signal can then be used later to make decisions (i.e.
for Zero-Touch such boards cannot participate, so even if someone
gets their hands on such a board and then write SN Bits to it, it
won't be able to be used for Zero-Touch).

BUG=b:111195266
TEST=TBD (manual test by louiscollard@ on sokara?)
BRANCH=None

Change-Id: I4ba0016e9a8485eee28fcdb083f74cf81caa0783