usb_pd: Retry sending VDMs on transmit failure.
When VDM messages are queued up to be sent and encounter a transmit
failure, there are no retry mechanisms in place. This leads to cases
where we think we had sent a reply but it wasn't actually sent and
violates the PD 2.0+ spec.

This commit simply attempts to retry sending the VDM when it encounters
an error.

BUG=b:117902978
BRANCH=firmware-nocturne-10984.B
TEST=Flash nocturne; Plug in Apple 87W USB-C charger, verify that
retries are attempted when a reply to DiscIdent fails to be transmitted.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/1288036
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Change-Id: I9c0ed392e27482e4e0b87cfa52b82941e3679eb5