usb_pd: Retry sending VDMs on transmit failure.
When VDM messages are queued up to be sent and encounter a transmit
failure, there are no retry mechanisms in place. This leads to cases
where we think we had sent a reply but it wasn't actually sent and
violates the PD 2.0+ spec.

This commit simply attempts to retry sending the VDM when it encounters
an error.

BUG=b:117902978
BRANCH=firmware-nocturne-10984.B
TEST=Flash nocturne; Plug in Apple 87W USB-C charger, verify that
retries are attempted when a reply to DiscIdent fails to be transmitted.

Change-Id: Ibddb82f7ba833cfed5d4843ab68fd57a2424ceb3
Signed-off-by: Aseda Aboagye <aaboagye@google.com>