drivers/accelgyro_lsm6dsm: Fix the FIFO_NUM_DEV configuration
Adding FIFO_DEV_MAG increases the number of motion sensors in the FIFO
data set. In the boards where magnetometer is not present, this causes
fifo_enable function to perform an out of bound memory access and lead
to a crash. Update the number of motion sensors with FIFO_DEV_MAG only
in the boards where magnetometer is present.

BRANCH=None
TEST=bootup to ChromeOS in grabbiter and meep.
BUG=b:118083418

Change-Id: I77ce059cb05b429eff84d69c29081c8042fc2c51
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/1293889
Reviewed-by: Jett Rink <jettrink@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>