it83xx/clock: turn off VCC before changing PLL
Symptom:
On Ampton board, EC's clock frequency is not correct after changing
PLL sequence.

We have added a workaround to fix changing PLL failure issue while
CS# pin's level is low, but it does not fix the issue on the DX version.
We need to turn off VCC power to fix this issue on the DX.

BUG=b:70537592
BRANCH=none
TEST=EC clock frequency is correct on DX version after changing PLL
     sequence.

Change-Id: Id2f507d6b15da40bab178f2754c7b11d64ff5ddf
Signed-off-by: Dino Li <Dino.Li@ite.com.tw>