touchpad_st: update logical maxmium of y value
To make X and Y have the same resolution (ratio between logical
coordinate and physical coordinate), the logical maxmium of y is changed
to 1573.

This change requires ST firmware v36.0 (0x24) or above.
If this is used on older ST firmware, it will still work, but the Y
movement will be scaled.

BRANCH=nocturne
BUG=b:118363768, b:119236031
TEST=make buildall -j64, test on whiskers
Signed-off-by: Wei-Han Chen <stimim@chromium.org>

Change-Id: Ic6c5db496c28a2f0febb8a4073d738cd9ba45087