cheza: Avoid wrong test pin sampling on EC
The signal PMIC_KPD_PWR_ODL is connected to EC GPIOD6 which is the JTAG0 Enable pin. EC is pulled
on 3.3V while PMIC is on 1.8V, which may lead to an undesired voltage level during startup and
potentially triggers EC into a test mode. Adding a series diode and reconfigure D6 as an output
pin should resolve this issue.

BRANCH=none
BUG=b:78910672
TEST=Tested GPIOD6 with series diode work properly both S5->S0/S0->S5 on an EVT board.

Change-Id: I7ac90d7c4f0c56353781223366bb9ce4e0f4d723