flash_ec: change serial number detection mechanism
serial number is detected by checking in servod for
servo_micro_for_$BOARD_serialname, however this only works
if the EC name and the servo overlay name match. Which they
often don't.

This change introduces a $BOARD serialname only for base
yoshi flex connectors, and uses the generic servo_micro_serialname
control otherwise.

It is totally unclear if this actually works in a config with
a base, but it does seem to work for phaser.

BRANCH=None
BUG=b:117342440
TEST=works on phaser, probably broken on everything else.

Signed-off-by: Nick Sanders <nsanders@chromium.org>

Change-Id: I535f5e7005fa47e029039da144c24cc14cf8ea4d