Octopus: Vbus gone from USB-C port after cold reboot with USB-A drive
When Vbus turns on while we're in the process of running
nx20p348x_vbus_source_enable(), a sink can incorrectly be detected as a
charger.  This change moves the initialization of the flag indicating
we're attempting to source Vbus, and will restore the previous flag
state on failure.

BRANCH=None
BUG=b:117616479
TEST=on yorp, cold booted with USB key and saw it was not detected as a
charger on both ports, cold booted with actual charger to verify it was
correctly detected on both ports

Change-Id: Ie8de18f4b4cd335118d5d44e1476b0ececbcc029
Signed-off-by: Diana Z <dzigterman@chromium.org>