iteflash: Refactor I2C interface handling.
Currently iteflash supports two I2C interfaces, CCD and FTDI.  While
tracked in an enum, the function calls were handled by many "if" branches.

This replaces the "if" branches after initial option parsing with a struct
of callbacks, which can be implemented for any I2C interface.

This replaces the -c / --ccd flag with a -c / --i2c-interface flag which
takes an argument of an interface name, currently "ccd" or "ftdi", the
latter being the default for compatibility with current flash_ec invocation
of iteflash.  (Refactoring bonus: The new flag can be specified anywhere on
the command line.)

This adds a -W / --send_waveform boolean flag which can be set explicitly
true or false.  The default is false, with flash_ec updated to set it to
true to maintain the current behavior.

This change is in preparation for supporting a third interface type, the
Linux i2c-dev interface to Linux kernel I2C adapters.  In the long term
direct use of FTDI or CCD I2C interfaces may become unused, however at the
very least there will need to be a migration period where all three I2C
interfaces are supported, and we may want to continue supporting direct use
of CCD and possibly FTDI indefinitely.  (Both CCD and FTDI / Servo v2 will
be supported indirectly via Linux i2c-dev, along with other Servo
versions.)

BRANCH=none
BUG=b:79684405
TEST=With Servo v2 "flash_ec --board=bip" continues to work.

Change-Id: I6d96777016d05947acc5b0c073a315dfa2fe72c3
Signed-off-by: Matthew Blecker <matthewb@chromium.org>