iteflash: Refactor I2C interface handling.
Currently iteflash supports two I2C interfaces, CCD and FTDI.  While
tracked in an enum, the function calls were handled by many "if" branches.
This replaces the "if" branches after initial option parsing with a struct
of callbacks, which can be implemented for any I2C interface.

This also changes the -c / --ccd flag to a -c / --i2c-interface flag which
takes an argument of an interface name (currently "ccd" or "ftdi").

This change is in preparation for supporting a third interface type, the
Linux i2c-dev interface to Linux kernel I2C adapters.  In the long term
direct use of FTDI or CCD I2C interfaces may become unused, however at the
very least there will need to be a migration period where all three I2C
interfaces are supported, and it is possible we will want to continue
supporting direct use of CCD and possibly FTDI indefinitely.  (Both CCD
and FTDI will be supported indirectly via Linux i2c-dev, among others.)

BRANCH=none
BUG=b:79684405
TEST=Compiles, and --help looks correct.  I do not have an ITE board handy
as I write this to verify either CCD or FTDI functionality with, so
verification help would be appreciated.

Change-Id: I6d96777016d05947acc5b0c073a315dfa2fe72c3
Signed-off-by: Matthew Blecker <matthewb@chromium.org>