CHERRY-PICK: nocturne: Use avg pin value for disconnection.
Previously, the base detection state machine would monitor the detach
pins for determining when the base appeared to be detached.  However,
sometimes the voltage sensed on that pin would be lower than the detach
threshold for 1-2 readings after debouncing.  This would lead to a brief
detach and reattach event.

This commit changes the debouncing logic to the following:

* Assume the base is attached.
* If a reading is sensed where the detach pin is less than the set
  threshold, start debouncing and collect samples at a much more rapid
  rate.
* Set a deadline in the future when we'll make our decision.
* Once the deadline has expired, take an average of the last 5 samples
  and see if it's beneath our threshold or not.  If it's beneath the
  threshold, we assume the base to be disconnected, otherwise it's still
  connected.

BUG=b:118213312
BRANCH=firmware-nocturne-10984.B
TEST=Flash nocturne; attach whiskers, verify that whiskers doesn't
appear to briefly detach.
TEST=Apply base power with whiskers disconnected.  Wait until nocturne
enters base attach state, verify that it transitions to the detach state
very quickly.
TEST=Attach and detach whiskers; verify that the detach is recognized
quickly.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/1300614
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>
Change-Id: Ibd2aa911c95d8701962867ed8988d9aa1d70f9fa