cheza: Remove the internal PU for UART pins
It is not needed; otherwise, it pulls the UART_DB_TX_EC_RX
to 1.8V and and may affect cr50 servo detection.

BRANCH=none
BUG=b:117296184
TEST=Measured the voltage level of UART_DBG_TX_EC_RX.
When servo plugged and CCD unplugged, it shows 3.3V.
When servo unplugged and CCD unplugged, it shows 0V (was 1.8V).
When servo unplugged and CCD plugged, it shows 3.3V.
EC console works over either servo and CCD.

Change-Id: Ie81d1de0075ea1b2409131471fe8571b5bfdccd1
Signed-off-by: Wai-Hong Tam <waihong@google.com>