Update patch set 1
Patch Set 1: Code-Review-1

But you need a pullup on this rail when servo and CCD is not connected. Otherwise noise on this line will cause spurious interrupts on the EC, potentially affecting EC power.

We either need an external pullup (to the right voltage) or we need to fix by changing the power domains of pins, maybe swapping pins around (which is gonna be fun with uart).

Patch-set: 1
Reviewer: Gerrit User 1113904 <1113904@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1