power/common: Wait some time before updating wake masks
On platforms like KBL, the device keeps waking up periodically
for very short intervals to allow some SoC components to do
book-keeping activities. If the user happens to trigger EC-related
wake event during this short window, then the wakeup event could be
missed because it looks like the host is in S0 to the EC.

In order to avoid the race condition, update wake mask using a
deferred call to allow the system state to stabilize.

BUG=b:118490626
BRANCH=nocturne
TEST=No more lid open failures observed on nocturne.

Change-Id: I13f9f5760aaf7e54c676f43c48f9fc8de572fd01
Signed-off-by: Furquan Shaikh <furquan@google.com>