Update patch set 1
Patch Set 1: -Code-Review

> There aren't any other signals cr50 can use to detect servo.

Sigh, fine.

> Is this causing issues on any boards?

Not that we know of, but it's bad practice leave a uart signal low. You're essentially sending a continuous stream of start bits.

> AFAIK power has been fine in the past.

I generally mean EC power here. It might be spinning on a bad symbol irq depending on the peripheral implementation. Better ones might be ok.

> If future ECs require some other behavior, we can add a new strapping option and figure out something to do for them.

It's an ok plan. It's just that if it's not that bad to figure out now pehaps we should. If it is complicated, fine, we can punt.

> Filed an issue to request Nuvoton to clarify the impact:

It's not just a nuvoton issue, could be a problem with any EC we ever used (or will ever use) with cr50.

Patch-set: 1
Reviewer: Gerrit User 1113904 <1113904@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0