Battery: Add API to get display charge percentage
This patch converts the actual battery percentage to the display
percentage using the same conversion used by the OS and ACPI. EC
can use this percentage to control LEDs synchronously to the value
on the display.

We have a battery policy that we consider everything above (currently)
97% of full_capacity as 'full'. It is for the following purposes and
advantages:

1 Avoid Rohm charger's high pitch noise at high charge.
2 Avoid showing discharge state while AC is plugged (because gas
  gauge cycles charge-discharge)
3 Detect full state of battery packs which don't update full_capacity
  (as battery ages).

We want a single place where this policy is enforced and it is ACPI
(battery.asl). ACPI converts every capacity reading above
full_capacity to full_capacity. Thus, the kernel, powerd, and all
other user space programs get 'full' state consistently. Of course,
they don't know how full the battery is. It could be anywhere between
97% and 100%.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b:109954565,b:80270446
BRANCH=none
TEST=Verify charge LED changes to white (full) on Sona synchronously
to the display percentage.
TEST=Verify charge LED changes to blinking white (low) on Sona
within 30 seconds synchronously to the display percentage.

Change-Id: I2041cb768dee27b8dba94a32db0eb62dfa14c73b