Update patch set 16
Patch Set 16:

(2 comments)

honestly, I don't understand why it is necessary to keep changing in the same patch, this could be a separate change.

Also, it is a bit late, but in general size variables should be of size_t type, this could be fixed separately.

And also, whitespace changes should be separate, it reduces the size of patches and makes it easier to review.

Patch-set: 16
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>