fpsensor: add TPM seed to Input Key Material
Make the encryption key also depend on data held by the TPM. Append that
seed to the anti-rollback IKM and feed that to HKDF-Extract when
deriving the encryption key.
The seed must be set once, and can't be overwritten.

Bump the template format version to 3, since it's not compatible with
previously enrolled templates.

Also add the corresponding command to ectool (fpseed).

BRANCH=nocturne
BUG=b:117909326
TEST=upload templates without having set a seed -> fails as expected
TEST=enroll finger without having set a seed -> fails as expected
TEST=set a seed twice -> 2nd time fails as expected
TEST=set seed, enroll finger -> success.
TEST=upload templates after having set the seed -> success.
TEST=set a different seed, upload templates -> fails as expected
TEST=reboot EC, reset original seed, upload templates -> success
TEST=load templates enrolled with format version=2 -> fails as expected

Change-Id: I64fd99f9d317d1fcab4a58a679be64cf8a425b00
Signed-off-by: Nicolas Norvez <norvez@chromium.org>