gsctool: add RO_B and RW_B back to the output of gsctool -b
Contents in FW image header slot A and B should be identical. Previously I
removed slot B from the output of gsctool -b to avoid redundancy, but that
change broke a cr50 test. Add the slot B output back to fix it but keep the
warning message of slot A and B not matching.

Note that if machine output is enabled (with option -M) and if slot A and B
are not identical, the header info won't be outputted and gsctool will
still terminate with a non-zero exit status.

BRANCH=none
BUG=b:118701324
TEST=manually run gsctool on a soraka device
     $ gsctool -b cr50.bin.prod
     RO_A:0.0.10 RW_A:0.3.10[00000000:00000000:00000000] RO_B:0.0.10
     RW_B:0.3.10[00000000:00000000:00000000]

Change-Id: I0d76052ef9b6e34c4c8f4be40ac48ed3559ef3ec
Signed-off-by: Wei-Cheng Xiao <garryxiao@chromium.org>