Update patch set 3
Patch Set 3:

one issue we faced with gcc8 is that it's link time optimization is more aggressive. If you have C code that is only referred to from assembly code, it might be optimized away. We had that issue on other platforms' USB handling tables, see https://chromium-review.googlesource.com/1177710 for example.

Patch-set: 3