Battery: Apply full factor to full capacity
This change applies 'full factor' to the battery full capacity. It
makes the rest of the system see consistent charge percentage behavior.

More concretely, with this change we can get rid of 'full factor'
from Powerd because it sees the current charge equal to the full
capacity.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b:109954565,b:80270446
BRANCH=none
TEST=Verify charge LED changes to white (full) on Sona synchronously
to the display percentage.
TEST=Verify charge LED changes to blinking white (low) on Sona
within 30 seconds synchronously to the display percentage.

Change-Id: Ifb27c802b0cf04195ac5b426c13f9476189feb75