Nami: Apply full factor to full capacity
This change applies 'full factor' to the battery full capacity. It
makes the rest of the system see consistent charge percentage behavior.

More concretely, with this change we can get rid of 'full factor'
from Powerd because it sees the current charge equal to the full
capacity.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

CQ-DEPEND=CL:1314048
BUG=b:109954565,b:80270446
BRANCH=none
TEST=Verify display percentages printed by EC and power_supply_info
move up synchronously on charge and the LED and the taskbar icon
turn to full at the same time.

Change-Id: Ic16463d457a6c5c2860b97476c78bdafb9021572