Battery: Set full_capacity charge host event threshold to zero
Currently, when a battery updates the full_capacity, a host is notified
only if it goes up or down more than 5 mAh. This workaround is not
needed for good batteries.

This change sets the threshold to zero except for Samus.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b:109954565,b:80270446
BRANCH=none
TEST=Verify display percentages printed by EC and power_supply_info
move up synchronously on charge and the LED and the taskbar icon
turn to full at the same time.

Change-Id: Ice0de712515e25929604376c8fe487b2be14b22c