Battery: Set full_capacity charge host event threshold to zero
Currently, when a battery updates the full_capacity, a host is notified
only if it goes up or down more than 5 mAh. This workaround is not
needed for good batteries. Actually, it's harmful for them because they
repeatedly update full_capacity in small steps. The gap becomes larger
and larger without the host notified.

This change sets the threshold to zero.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b:109954565,chromium:899120
BRANCH=none
TEST=Verify LED changes from amber (charging) to white (full) at the
same time as the display shows battery is full.

Change-Id: Ice0de712515e25929604376c8fe487b2be14b22c