Grunt: Advertise power availability in PDOs
PD Source PDO's have the capability to advertise when the port is
externally powered.  Advertise that fact for Grunt when the charge
manager has enough power available to clear the Low Power Charger
warning.

BUG=b:118597992, b:118846062
TEST=Connect two careenas together, where one of them is externally
powered, and observe that the non-powered device does not attempt to
role swap to become a source after negotiation.
BRANCH=firmware-grunt

Change-Id: I914dd30d93f8248eeebc9b98b7fe0ae4e9bb460e
Signed-off-by: Jonathan Brandmeyer <jbrandmeyer@chromium.org>