Grunt: Respect the charger override when role swapping
The user's request in the Power Settings dialog to charge from a
dual-role type-C device is communicated to the EC via the
charge_manager's override variable.  Respect that choice when deciding
whether or not to perform a role swap.

Additionally, do not perform a role swap to become a sink from a
dual-role device when we have external power available.  This policy
should be sufficient to Just Work (TM) when presented with a type-C
rechargeable battery pack.

TEST=none
BRANCH=firmware-grunt
BUG=b:118846062, b:118597992
Signed-off-by: Jonathan Brandmeyer <jbrandmeyer@chromium.org>

Change-Id: If5793617805de8b179c4f1c2bf6d4bb718e89386