core/minute-ia: Reset Task-switched flag during init
It is observed that CRO.TS may be set prior to main
ISH FW entry. In such case, an ESC instruction such
as fninit assembly instruction for FPU initialization,
will cause a fault.

The execution handlers are not setup this early in
FW initialization phase, and will lead to ISH CPU
reset.

This patch resets the TS flag in early FW initilization.

BRANCH=none
BUG=b:120051132
TEST=Test host FW loading for main ISH FW.

Change-Id: I7e330e2d1f39cc4a349f308ec0d046c19db281de
Signed-off-by: Rushikesh S Kadam <rushikesh.s.kadam@intel.com>