Update patch set 4
Patch Set 4:

It looks like this CL is currently targeting the firmware-grunt-11031.B branch directly, with the intent to cherry-pick it to the master branch later.  Please reverse the order, such that this change applies to master first, and can be cherry-picked to the firmware branch later.

Patch-set: 4
Reviewer: Gerrit User 1284593 <1284593@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>