sn5s330: Add VCONN overcurrent logging.
The TI SN5S330 has VCONN overcurrent protection and will automatically
latch off VCONN if a port overcurrents it.  Keeping VCONN latched off is
allowed per the USB Type-C spec.  This commit simply logs a message on
the EC console such that it will be easier to tell when it happens.

BUG=none
BRANCH=firmware-nocturne-10984.B
TEST=flash nocturne, plug in a Type-C adapter which overcurrents VCONN,
verify that a console message indicating the overcurrent event is shown.

Change-Id: I22da6e8d43ac7739dc213d5a1f049b98d7ee829d
Signed-off-by: Aseda Aboagye <aaboagye@google.com>