sn5s330: Add VCONN overcurrent logging.
The TI SN5S330 has VCONN overcurrent protection and will automatically
latch off VCONN if a port overcurrents it.  Keeping VCONN latched off is
allowed per the USB Type-C spec.  This commit simply logs a message on
the EC console such that it will be easier to tell when it happens.

BUG=none
BRANCH=firmware-nocturne-10984.B
TEST=flash nocturne, plug in a Type-C adapter which overcurrents VCONN,
verify that a console message indicating the overcurrent event is shown.

Change-Id: I55c6acec656f5807e81906684678233eb2689e65
Signed-off-by: Aseda Aboagye <aaboagye@google.com>