Update patch set 1
Patch Set 1:

I'm testing with the PS8805. Here's a case to try out:

* Plug in C to A adapter while in S0.
* Suspend the DUT.
* Unplug the C to A adapter.
* Plug in a charger to that same port.

Expected: source is detected and device starts charging.
Actual: Nothing is detected and device does not charge.

Then plug in the C to A adapter back in to the port and unplug. Then plug in a charger, and the source is detected.

Yes, the drp_state is updated so it doesn't toggle when the sink is removed, but also the Rp's remain asserted. Also, for some reason the API to disable DRP was removed. (IIRC because it was no longer being used).

Patch-set: 1
Reviewer: Gerrit User 1125984 <1125984@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>