USB-PD: Supply power up to 3A + 1.5A
Currently, USB PD ports supply 3A only if there is no other active
supplier. We enforce this rule even if the port is actively supplying
power. That is, we drop the max current of an active port to 1.5A if
a sink device is plugged to another port.

This change makes USB PD ports supply 3A if the other ports are not
supplying 3A.

(P0, P1) and '*' indicates a sink device is plugged.

  Unplug both: (3A, 3A)

  Plug P0: (*3A, 1.5A)
  Plug P1: (*3A, *1.5A)
  Unplug P0: (1.5A, *3A)
  Unplug P1: (3A, 3A)

  Plug P1: (1.5A, *3A)
  Plug P0: (*1.5A, *3A)
  Unplug P0: (1.5A, *3A)
  Unplug P1: (3A, 3A)

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b:115291657
BRANCH=none
TEST=Performed as shown above and verify current of active port is not
affected by the other port.

Change-Id: I08fb04da7e0177d5e71f823fb1e47e6945ae12fc
Reviewed-on: https://chromium-review.googlesource.com/1322069
Commit-Ready: Daisuke Nojiri <dnojiri@chromium.org>
Tested-by: Raymond Chou <raymond_chou@compal.corp-partner.google.com>
Reviewed-by: Jett Rink <jettrink@chromium.org>