chip/npcx: Default output the i2c_recovery() error message
The i2c_recovery() will be called if the i2c bus error is detected.
This is an serious error handling and may be caused by the slave
which doesn't follow the spec. We should output this error message
to let developer catch this error at the beginning like proto stage.

BUG=b:118063849
BRANCH=master
TEST=error message popped out on Nami and Rammus

Change-Id: I4996cd18415d6ee4c5cd48ac374252a2230628b0
Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>