Update patch set 15
Patch Set 15:

> Patch Set 15:
> 
> > Patch Set 15:
> > 
> > > Patch Set 15:
> > > 
> > > Actually, it already does what we mean by absolute window check. The host memmap holds the pivot value.
> > 
> > This seems like a relative window pivot. I.e. the window is always based on the current value (instead of the last time the event was sent)
> 
> The pivot value (memmap_lfcc) doesn't change until the full capacity goes over a threshold.
> 
> 	if (...
> 	    (curr.batt.full_capacity <= (*memmap_lfcc - LFCC_EVENT_THRESH) ||
> 	     curr.batt.full_capacity >= (*memmap_lfcc + LFCC_EVENT_THRESH))) {
> 		*memmap_lfcc = curr.batt.full_capacity;
> 		...
> 	}

I seem to be extra slow on the CL :( I am surprise that the absolute window didn't work. If we change the default threshold to 1 or 2, does it not meet the needs of the constantly updating battery (new scenario). If we chose 2, then we would only be off by 5mA at most.

If a tighter default threshold doesn't work, then I am fine with this CL as is

Patch-set: 15
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>