Battery: Use host full capacity to compute display percentage
The full capacity known by the host may slightly differ from the full
capacity known by the EC because we notify the host of the new
capacity only if the difference is larger than 5 mAh.

This patch makes EC use the host full capacity instead of the local
full capacity to compute the display percentage.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b:109954565,b:80270446
BRANCH=none
TEST=Verify display percentages printed by EC and power_supply_info
move up synchronously on charge and the LED and the taskbar icon
turn to full at the same time.

Change-Id: Ie695a9937a22fc7a769b82448f4600d4491935b3