PRESUBMIT: sanity check include/ec_commands.h for kernel-doc
.../kernel-doc -none /path/to/file

will do the kernel-doc parsing, only emitting errors. Run it only on the
ec_commands.h header, since that's often synced with the Linux sources.

I make a small adjustment to the kernel-doc script, so it can accept
sticking the __ec_align(|_size|_offset)[124] attributes in the "wrong"
place. We can work around that when syncing with a Linux tree.

BRANCH=none
BUG=chromium:621123
TEST=run the presubmits

Change-Id: I56645e613f8689c26ddf6ded1325e270be35da8a
Signed-off-by: Brian Norris <briannorris@chromium.org>